feat: 카프카를 통해 상품 가격도 전달

+++ apps/commerce-streamer/src/main/java/com/loopers/application/metrics/MetricsStrategy.java
@@ -14,7 +14,7 @@ public abstract class MetricsStrategy {
 
   abstract MetricsMethod method();
 
-  public void increment(Long productId, double weight, int value) {
+  public void increment(Long productId, double weight, long value) {
     rankingRepository.increment(productId, weight * value);
   }
 }
